--- a/benchmarks/CMakeLists.txt
+++ b/benchmarks/CMakeLists.txt
@@ -25,6 +25,8 @@ endif()
 
 # Build a common, helper library.
 add_library( vecmem_benchmark_common STATIC
+   "common/make_jagged_sizes.hpp"
+   "common/make_jagged_sizes.cpp"
    "common/make_jagged_vector.hpp"
    "common/make_jagged_vector.cpp" )
 target_link_libraries( vecmem_benchmark_common
@@ -34,4 +36,7 @@ target_link_libraries( vecmem_benchmark_common
 add_subdirectory(core)
 if(VECMEM_BUILD_CUDA_LIBRARY)
    add_subdirectory(cuda)
+endif()
+if(VECMEM_BUILD_SYCL_LIBRARY)
+   add_subdirectory(sycl)
 endif()