--- a/benchmarks/common/make_jagged_vector.hpp
+++ b/benchmarks/common/make_jagged_vector.hpp
@@ -14,6 +14,7 @@
 
 // System include(s).
 #include <cstddef>
+#include <vector>
 
 namespace vecmem::benchmark {
 
@@ -22,14 +23,11 @@ namespace vecmem::benchmark {
 /// It creates a jagged vector with a fixed "outer size", and random sized
 /// "inner vectors" that would not be larger than some specified value.
 ///
-/// @param outerSize The fixed "outer size" of the resulting vector
-/// @param maxInnerSize The maximum for the random "inner sizes" of the
-///                     resulting vector
+/// @param sizes The sizes of the vectors in the jagged vector
 /// @param mr The memory resource to use
 /// @return A jagged vector with the specifier properties
 ///
-jagged_vector<int> make_jagged_vector(std::size_t outerSize,
-                                      std::size_t maxInnerSize,
+jagged_vector<int> make_jagged_vector(const std::vector<std::size_t>& sizes,
                                       memory_resource& mr);
 
 }  // namespace vecmem::benchmark