--- a/benchmarks/core/benchmark_copy.cpp
+++ b/benchmarks/core/benchmark_copy.cpp
@@ -11,12 +11,14 @@
 #include <vecmem/utils/copy.hpp>
 
 // Common benchmark include(s).
+#include "../common/make_jagged_sizes.hpp"
 #include "../common/make_jagged_vector.hpp"
 
 // Google benchmark include(s).
 #include <benchmark/benchmark.h>
 
 // System include(s).
+#include <numeric>
 #include <vector>
 
 namespace vecmem::benchmark {
@@ -26,23 +28,132 @@ static host_memory_resource host_mr;
 /// The copy object to use in the benchmark(s).
 static copy host_copy;
 
-/// Function benchmarking the @c vecmem::copy jagged vector operations
-void jaggedVectorHostCopy(::benchmark::State& state) {
+/// Function benchmarking "unknown" host-to-device jagged vector copies
+void jaggedVectorUnknownHtoDCopy(::benchmark::State& state) {
+
+    // Generate the sizes of the jagged vector/buffer for the test.
+    const std::vector<std::size_t> sizes =
+        make_jagged_sizes(state.range(0), state.range(1));
+
+    // Set custom "counters" for the benchmark.
+    const std::size_t bytes = std::accumulate(sizes.begin(), sizes.end(),
+                                              static_cast<std::size_t>(0u)) *
+                              sizeof(int);
+    state.counters["Bytes"] = static_cast<double>(bytes);
+    state.counters["Rate"] =
+        ::benchmark::Counter(static_cast<double>(bytes),
+                             ::benchmark::Counter::kIsIterationInvariantRate,
+                             ::benchmark::Counter::kIs1024);
 
     // Create the "source vector".
-    jagged_vector<int> source =
-        make_jagged_vector(state.range(0), state.range(1), host_mr);
+    jagged_vector<int> source = make_jagged_vector(sizes, host_mr);
     const data::jagged_vector_data<int> source_data = get_data(source);
-    // Create the "destination vector".
-    jagged_vector<int> dest;
+    // Create the "destination buffer".
+    data::jagged_vector_buffer<int> dest(sizes, host_mr);
+    host_copy.setup(dest);
 
     // Perform the copy benchmark.
     for (auto _ : state) {
-        dest.clear();
         host_copy(source_data, dest);
     }
 }
 // Set up the benchmark.
-BENCHMARK(jaggedVectorHostCopy)->Ranges({{10, 100000}, {50, 5000}});
+BENCHMARK(jaggedVectorUnknownHtoDCopy)->Ranges({{10, 100000}, {50, 5000}});
+
+/// Function benchmarking "known" host-to-device jagged vector copies
+void jaggedVectorKnownHtoDCopy(::benchmark::State& state) {
+
+    // Generate the sizes of the jagged vector/buffer for the test.
+    const std::vector<std::size_t> sizes =
+        make_jagged_sizes(state.range(0), state.range(1));
+
+    // Set custom "counters" for the benchmark.
+    const std::size_t bytes = std::accumulate(sizes.begin(), sizes.end(),
+                                              static_cast<std::size_t>(0u)) *
+                              sizeof(int);
+    state.counters["Bytes"] = static_cast<double>(bytes);
+    state.counters["Rate"] =
+        ::benchmark::Counter(static_cast<double>(bytes),
+                             ::benchmark::Counter::kIsIterationInvariantRate,
+                             ::benchmark::Counter::kIs1024);
+
+    // Create the "source vector".
+    jagged_vector<int> source = make_jagged_vector(sizes, host_mr);
+    const data::jagged_vector_data<int> source_data = get_data(source);
+    // Create the "destination buffer".
+    data::jagged_vector_buffer<int> dest(sizes, host_mr);
+    host_copy.setup(dest);
+
+    // Perform the copy benchmark.
+    for (auto _ : state) {
+        host_copy(source_data, dest, copy::type::host_to_device);
+    }
+}
+// Set up the benchmark.
+BENCHMARK(jaggedVectorKnownHtoDCopy)->Ranges({{10, 100000}, {50, 5000}});
+
+/// Function benchmarking "unknown" device-to-host jagged vector copies
+void jaggedVectorUnknownDtoHCopy(::benchmark::State& state) {
+
+    // Generate the sizes of the jagged vector/buffer for the test.
+    const std::vector<std::size_t> sizes =
+        make_jagged_sizes(state.range(0), state.range(1));
+
+    // Set custom "counters" for the benchmark.
+    const std::size_t bytes = std::accumulate(sizes.begin(), sizes.end(),
+                                              static_cast<std::size_t>(0u)) *
+                              sizeof(int);
+    state.counters["Bytes"] = static_cast<double>(bytes);
+    state.counters["Rate"] =
+        ::benchmark::Counter(static_cast<double>(bytes),
+                             ::benchmark::Counter::kIsIterationInvariantRate,
+                             ::benchmark::Counter::kIs1024);
+
+    // Create the "source buffer".
+    data::jagged_vector_buffer<int> source(sizes, host_mr);
+    host_copy.setup(source);
+    // Create the "destination vector".
+    jagged_vector<int> dest = make_jagged_vector(sizes, host_mr);
+    data::jagged_vector_data<int> dest_data = get_data(dest);
+
+    // Perform the copy benchmark.
+    for (auto _ : state) {
+        host_copy(source, dest_data);
+    }
+}
+// Set up the benchmark.
+BENCHMARK(jaggedVectorUnknownDtoHCopy)->Ranges({{10, 100000}, {50, 5000}});
+
+/// Function benchmarking "known" device-to-host jagged vector copies
+void jaggedVectorKnownDtoHCopy(::benchmark::State& state) {
+
+    // Generate the sizes of the jagged vector/buffer for the test.
+    const std::vector<std::size_t> sizes =
+        make_jagged_sizes(state.range(0), state.range(1));
+
+    // Set custom "counters" for the benchmark.
+    const std::size_t bytes = std::accumulate(sizes.begin(), sizes.end(),
+                                              static_cast<std::size_t>(0u)) *
+                              sizeof(int);
+    state.counters["Bytes"] = static_cast<double>(bytes);
+    state.counters["Rate"] =
+        ::benchmark::Counter(static_cast<double>(bytes),
+                             ::benchmark::Counter::kIsIterationInvariantRate,
+                             ::benchmark::Counter::kIs1024);
+
+    // Create the "source buffer".
+    data::jagged_vector_buffer<int> source(sizes, host_mr);
+    host_copy.setup(source);
+    // Create the "destination vector".
+    jagged_vector<int> dest = make_jagged_vector(sizes, host_mr);
+    data::jagged_vector_data<int> dest_data = get_data(dest);
+
+    // Perform the copy benchmark.
+    for (auto _ : state) {
+        host_copy(source, dest_data, copy::type::device_to_host);
+    }
+}
+// Set up the benchmark.
+BENCHMARK(jaggedVectorKnownDtoHCopy)->Ranges({{10, 100000}, {50, 5000}});
 
 }  // namespace vecmem::benchmark